Add files via upload
 follow the hedging strategy of shorting a bull spread using Neural Networks (ANNs) compared
with the Black Scholes model. The strategy is to be implemented such that the Profit-loss achieves
a conditional Value-at-Risk no greater than $-0.02. The ANN’s performance increases with structural
changes by increasing the hidden layers and nodes which becomes similar to the Black Scholes. The
Recurrent Neural Network outperforms both the best ANN and the Black Scholes model. This criteria
is given by the smallest price adjustment which needs to be made the financial derivative to achieve the
strategy